requirements.txt
aioredis==1.3.1
asgiref==3.5.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
certifi==2022.6.15
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.1
charset-normalizer==2.1.1
constantly==15.1.0
coverage==6.4.4
cryptography==37.0.4
daphne==3.0.2
defusedxml==0.7.1
Django==4.1
django-allauth==0.51.0
django-cors-headers==3.13.0
django-rest-auth==0.9.5
djangorestframework==3.13.1
h11==0.13.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
msgpack==1.0.4
oauthlib==3.2.1
outcome==1.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.5.0
pyOpenSSL==22.0.0
PySocks==1.7.1
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
selenium==4.4.3
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.2
trio==0.21.0
trio-websocket==0.9.2
Twisted==22.4.0
txaio==22.2.1
types-cryptography==3.3.23
typing_extensions==4.3.0
urllib3==1.26.12
wsproto==1.2.0
zope.interface==5.4.0
